requirements.txt
psycopg2
python-dotenv
autopep8
django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
psycopg2-binary
celery[redis]
gevent
requests
django-celery-beat
drf-yasg
whitenoise